Fixes and improvements
Fixes:
creation of unnecessary row at line 157 of Arena.java. This error
occurred when multiple bubbles land on the last row consecutively.

The collision detection now stops when one collision is detected instead
of detecting multiple collision with the same object.

Improvements:
The addBubbleRow after 10 shot is removed from the Game.java since this
is already done inside the landBubble() of Arena.java.
The shot counter is removed from the Cannon class as well. This is not
needed anymore.